triangular tesselation
